pyproject.toml
[project]
name = "flaskproj"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "Flask",
]

[project.optional-dependencies]
dev = [
    "pytest-flask",
    "beautifulsoup4",
]
